requirements.txt
datasets==2.21.0
numpy==1.26.4
pandas==1.5.3
PyYAML==6.0.1
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.14.0
tokenizers==0.20.1
torch==2.6.0
tqdm==4.66.4
transformers==4.45.2
wandb==0.18.7